Use size_t for size in reg_aggregate_size of aarch64.

+++ c2mir/aarch64/caarch64-ABI-code.c
@@ -13,10 +13,10 @@ static int target_return_by_addr_p (c2m_ctx_t c2m_ctx, struct type *ret_type) {
 }
 
 static int reg_aggregate_size (c2m_ctx_t c2m_ctx, struct type *type) {
-  int size;
+  size_t size;
 
   if (type->mode != TM_STRUCT && type->mode != TM_UNION) return -1;
-  return (size = type_size (c2m_ctx, type)) <= 2 * 8 ? size : -1;
+  return (size = type_size (c2m_ctx, type)) <= 2 * 8 ? (int) size : -1;
 }
 
 static void target_add_res_proto (c2m_ctx_t c2m_ctx, struct type *ret_type,
